完善功能

+++ src/com/zhidi/servlet/taxer/DeleteTaxerServlet.java
@@ -31,9 +31,26 @@ public class DeleteTaxerServlet extends HttpServlet{
         ResponseResult result = new ResponseResult();
         Integer id = StringUtils.isEmpty(req.getParameter("id")) ? null : Integer.parseInt(req.getParameter("id"));
         TaxerDaoImpl taxerDao = new TaxerDaoImpl();
-        boolean state = taxerDao.delete(id);
-        result.setSuccess(true);
-        result.setMsg("删除成功");
+
+        //查询tb_tax_source是否有对应的子记录，如果有对应的子记录就不能删除，就返回一个提示信息：此办税专员有对应的任务，不能删除
+        boolean isHave = taxerDao.isHaveTaxOrgan(id);
+
+        if (!isHave) {
+            //调用删除的方法
+            boolean state = taxerDao.delete(id);
+            //设置响应信息
+            if (state) {
+                result.setSuccess(true);
+                result.setMsg("删除成功");
+            } else {
+                result.setSuccess(false);
+                result.setMsg("删除失败");
+            }
+        } else {
+            result.setSuccess(false);
+            result.setMsg("存在对应子记录，不能删除");
+        }
+
         JSONObject js = JSONObject.fromObject(result);
         PrintWriter writer = res.getWriter();
         writer.print(js.toString());
